requirements.txt
asgiref==3.2.10
Django==3.0.8
et-xmlfile==1.0.1
export==0.1.2
jdcal==1.4.1
matplotlib==3.3.0
mysqlclient==2.0.1
numpy==1.19.1
openpyxl==3.0.4
Pillow==7.2.0
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
workbook==1.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
